+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Moved search index building and seach GC jobs into `analyzer` instance.
 
 ## `20231018t131100-all`
  * Bumped runtimeVersion to `2023.10.18`.